+++ core/src/main/java/org/jdbi/v3/core/mapper/ColumnMappers.java
@@ -39,6 +39,7 @@ public class ColumnMappers implements JdbiConfig<ColumnMappers> {
         factories.add(new PrimitiveMapperFactory());
         factories.add(new BoxedMapperFactory());
         factories.add(new EssentialsMapperFactory());
+        factories.add(new InternetMapperFactory());
         factories.add(new BuiltInMapperFactory());
         factories.add(new SqlArrayMapperFactory());
     }
